feat(add-new-chain): seperate file for chain data
RE #KRYS-2734
