+++ KyberNetwork/KyberNetwork/Coordinators/BalanceCoordinator/KNLoadBalanceCoordinator.swift
@@ -363,22 +363,8 @@ class KNLoadBalanceCoordinator {
         completion(true)
       } else {
         var summaryChains: [KNSummaryChainModel] = []
-        if KNEnvironment.default == .ropsten {
-          summaryChains = [KNSummaryChainModel.defaultValue(chainId: Constants.ethRoptenPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.bscRoptenPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.polygonRoptenPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.avalancheRoptenPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.fantomMainnetRPC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.cronosMainnetRPC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.arbitrumMainnetRPC.chainID)]
-        } else {
-          summaryChains = [KNSummaryChainModel.defaultValue(chainId: Constants.ethMainnetPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.bscMainnetPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.polygonMainnetPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.avalancheMainnetPRC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.fantomMainnetRPC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.cronosMainnetRPC.chainID),
-                           KNSummaryChainModel.defaultValue(chainId: Constants.arbitrumMainnetRPC.chainID)]
+        ChainType.allCases.forEach { chain in
+          summaryChains.append(KNSummaryChainModel.defaultValue(chainId: chain.getChainId()))
         }
         BalanceStorage.shared.saveSummaryChainModels(summaryChains)
         KNNotificationUtil.postNotification(for: kOtherBalanceDidUpdateNotificationKey)
