+++ KyberNetwork/KyberNetwork/Screens/Tabbars/KyberSwap/NewDesign/ViewModel/KSwapViewModel.swift
@@ -890,44 +890,22 @@ class KSwapViewModel {
 
 extension KSwapViewModel {
   func getChain(chainId: Int) -> ChainType? {
-    switch chainId {
-    case Constants.ethMainnetPRC.chainID, Constants.ethRoptenPRC.chainID:
-      return .eth
-    case Constants.bscMainnetPRC.chainID, Constants.bscRoptenPRC.chainID:
-      return .bsc
-    case Constants.polygonMainnetPRC.chainID, Constants.polygonRoptenPRC.chainID:
-      return .polygon
-    case Constants.avalancheMainnetPRC.chainID, Constants.avalancheRoptenPRC.chainID:
-      return .avalanche
-    case Constants.fantomMainnetRPC.chainID:
-      return .fantom
-    case Constants.cronosMainnetRPC.chainID:
-      return .cronos
-    case Constants.arbitrumMainnetRPC.chainID:
-      return .arbitrum
-    default:
-      return nil
+    var chainType: ChainType?
+    ChainType.allCases.forEach { chain in
+      if chain.getChainId() == chainId {
+        chainType = chain
+      }
     }
+    return chainType
   }
 
   func chainName(chainId: Int) -> String? {
-    switch chainId {
-    case Constants.ethMainnetPRC.chainID, Constants.ethRoptenPRC.chainID:
-        return "Ethereum"
-    case Constants.bscMainnetPRC.chainID, Constants.bscRoptenPRC.chainID:
-        return "BSC"
-    case Constants.polygonMainnetPRC.chainID, Constants.polygonRoptenPRC.chainID:
-        return "Polygon"
-    case Constants.avalancheMainnetPRC.chainID, Constants.avalancheRoptenPRC.chainID:
-        return "Avalanche"
-    case Constants.fantomMainnetRPC.chainID:
-      return "Fantom"
-    case Constants.cronosMainnetRPC.chainID:
-      return "Cronos"
-    case Constants.arbitrumMainnetRPC.chainID:
-      return "Arbitrum"
-    default:
-        return nil
+    var chainType: ChainType?
+    ChainType.allCases.forEach { chain in
+      if chain.getChainId() == chainId {
+        chainType = chain
+      }
     }
+    return chainType?.chainName()
   }
 }
