+++ KyberNetwork/Tokens/Types/TokenData.swift
@@ -48,23 +48,23 @@ class Token: Codable, Equatable, Hashable {
   }
   
   var isBNB: Bool {
-    return self.address.lowercased() == Constants.bnbAddress.lowercased()
+    return self.address.lowercased() == AllChains.bscMainnetPRC.quoteTokenAddress.lowercased()
   }
 
   var isMatic: Bool {
-    return self.address.lowercased() == Constants.maticAddress.lowercased()
+    return self.address.lowercased() == AllChains.polygonMainnetPRC.quoteTokenAddress.lowercased()
   }
 
   var isAvax: Bool {
-    return self.address.lowercased() == Constants.avaxAddress.lowercased()
+    return self.address.lowercased() == AllChains.avalancheMainnetPRC.quoteTokenAddress.lowercased()
   }
 
   var isCro: Bool {
-    return self.address.lowercased() == Constants.cronosAddress.lowercased()
+    return self.address.lowercased() == AllChains.cronosMainnetRPC.quoteTokenAddress.lowercased()
   }
 
   var isFtm: Bool {
-    return self.address.lowercased() == Constants.fantomAddress.lowercased()
+    return self.address.lowercased() == AllChains.fantomMainnetRPC.quoteTokenAddress.lowercased()
   }
 
   var isQuoteToken: Bool {
